Upgrade efficiency of ShardManager#getGuildById (#378)
While looking at the additions, I realized that the ShardManager#getGuildById method can be more efficient.
Rather than finding the snowflake in a combined collection, you could calculate the shard id of the Guild, get the shard and obtain the guild from there instead. This upgrades the method from O(n) to O(1).